requirements.txt
# Core dependencies
langchain==0.1.12
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-community==0.0.25
openai==1.12.0

# Pydantic and Pydantic AI for agent building
pydantic==2.10.6
pydantic-core==2.27.2

# Vector storage and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.4.0

# Caching and optimization
tiktoken==0.9.0
tqdm==4.66.2
numpy==1.26.4

# Document processing
python-docx==1.0.1
PyPDF2==3.0.1

# Text processing and cleaning
clean-text==0.6.0
Unidecode==1.3.8

# Web scraping and news retrieval
beautifulsoup4==4.12.2
requests==2.31.0
duckduckgo-search==3.9.3
aiohttp==3.9.1

# Data analysis and visualization
pandas==2.2.0
matplotlib==3.10.1
seaborn==0.13.2
plotly==5.20.0
scikit-learn==1.6.1
scipy==1.15.2

# NLP and text analysis
nltk==3.8.1

# Report generation
Markdown==3.6
Jinja2==3.1.6
MarkupSafe==2.1.5

# Environment and configuration
python-dotenv==1.0.0

# Agent Workflows and Graph Orchestration
langgraph==0.3.9

# Web Framework and API
Flask==3.0.0
flask-cors==4.0.0
flask-session==0.5.0
flask-limiter==3.5.0
flask-restx==1.3.0
flask-caching==2.1.0
Werkzeug==3.0.1

# Additional dependencies
typing-extensions==4.12.2
charset-normalizer==3.4.1
certifi==2025.1.31
idna==3.10
urllib3==2.3.0
pypdf==5.4.0

# Rate limiting and retry logic
ratelimit==2.2.1
# New additions from the code block
gunicorn==21.2.0

